+++ app/services/discourse_community_platform/feeds/popular_topics.rb
@@ -72,7 +72,11 @@ module ::DiscourseCommunityPlatform
           end.first(@limit)
 
         scores = TopicScore.where(topic_id: topics.map(&:id)).index_by(&:topic_id)
-        communities = Community.where(category_id: topics.map(&:category_id)).index_by(&:category_id)
+        communities =
+          Community
+            .where(category_id: topics.map(&:category_id))
+            .includes(:icon_upload)
+            .index_by(&:category_id)
         votes = user_votes(topics)
         previews = TopicPreviews.call(topics:, guardian: @guardian)
 
@@ -99,17 +103,26 @@ module ::DiscourseCommunityPlatform
             user_vote: votes.fetch(topic.id, 0),
             excerpt: preview[:excerpt],
             image_url: preview[:image_url],
-            community: {
-              id: community.id,
-              name: community.name,
-              slug: community.slug,
-              path: "/s/#{community.slug}",
-            },
+            community: community_identity(community),
           }
         end
       end
 
       private
+
+      def community_identity(community)
+        icon_upload = community.icon_upload
+        icon_url = icon_upload&.url if icon_upload.blank? || @guardian.can_see_upload?(icon_upload)
+
+        {
+          id: community.id,
+          name: community.name,
+          slug: community.slug,
+          path: "/s/#{community.slug}",
+          icon_emoji: community.icon_emoji,
+          icon_url:,
+        }
+      end
 
       def user_votes(topics)
         user = @guardian.user